moves appropriate tests from enigma_test to shift_gen test
